+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 0.23.3 (provisional)
+## 0.24.0 (provisional)
 
 - Refactoring internal edge & neighbor iteration schemes.
 - Dropping undocumented `#.upgradeToMixed` & `#.upgradeToMulti`.