+++ src/types/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.24.0 (provisional)
+
+- Improving event emitter types.
+- Aligning type declarations to `graphology` 0.24.0.
+
 ## 0.23.0
 
 - Aligning type declarations to `graphology` 0.23.0.